Reconstruct party initialization, inventory and matrix animation

+++ include/game/equipment.h
@@ -3,5 +3,5 @@
 #include <game/save_data.h>
 
 void Equipment_GetBonuses(u16 equipment, s16 *hp, s16 *power, s16 *defense, s16 *speed, s16 *stache);
-void Equipment_Apply(int party_member, u16 equipment);
+void Equipment_Apply(u8 party_member, u16 equipment);
 #endif
